+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [1.7.1](https://github.com/GodIsI/homebridge-ups-monitor/compare/v1.7.0...v1.7.1) (2026-06-01)
+
+### Features
+
+* reclaim history/CSV logs from previous storage location ([271af8f](https://github.com/GodIsI/homebridge-ups-monitor/commit/271af8fb86eb574384ed6746d2d29a2124e45f09))
+
+### Bug Fixes
+
+* resolve storage path via api.user.storagePath() ([8525d4a](https://github.com/GodIsI/homebridge-ups-monitor/commit/8525d4aedd0163dfe493589c6bded90fc5a9a341))
+
 ## [1.7.0](https://github.com/GodIsI/homebridge-ups-monitor/compare/v1.6.1-beta.203...v1.7.0) (2026-06-01)
 
 ### Bug Fixes